Fix `Margin` not being accessible
